refactor(scancode): Split the model into classes in a separate package
This improves the overview quite a bit.
