+++ plugins/scanners/scancode/src/main/kotlin/model/CopyrightEntry.kt
@@ -1,0 +1,54 @@
+/*
+ * Copyright (C) 2023 The ORT Project Authors (see <https://github.com/oss-review-toolkit/ort/blob/main/NOTICE>)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * License-Filename: LICENSE
+ */
+
+package org.ossreviewtoolkit.plugins.scanners.scancode.model
+
+import kotlinx.serialization.Serializable
+
+/**
+ * An interface to be able to treat all versions of copyright entries the same.
+ *
+ * Note that the data class constructors of the individual version's implementation of this interface should contain
+ * only those properties which are actually present in the data, in the order used in the data, to exactly resemble that
+ * version's data model. Other properties required to implement the interface should be added to the body of the data
+ * class to clearly separate data model properties from "synthetic" interface properties.
+ */
+sealed interface CopyrightEntry {
+    val statement: String
+    val startLine: Int
+    val endLine: Int
+
+    @Serializable
+    data class Version1(
+        val value: String,
+        override val startLine: Int,
+        override val endLine: Int
+    ) : CopyrightEntry {
+        override val statement = value
+    }
+
+    @Serializable
+    data class Version2(
+        val copyright: String,
+        override val startLine: Int,
+        override val endLine: Int
+    ) : CopyrightEntry {
+        override val statement = copyright
+    }
+}
