+++ plugins/scanners/scancode/src/main/kotlin/model/FileEntry.kt
@@ -17,54 +17,19 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.plugins.scanners.scancode
+package org.ossreviewtoolkit.plugins.scanners.scancode.model
 
-import java.io.File
+import kotlin.collections.flatMap
+import kotlin.collections.orEmpty
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerializationException
-import kotlinx.serialization.json.JsonArray
-import kotlinx.serialization.json.JsonElement
-import kotlinx.serialization.json.JsonPrimitive
-import kotlinx.serialization.json.jsonPrimitive
 
+import org.ossreviewtoolkit.plugins.scanners.scancode.ScanCode
 import org.ossreviewtoolkit.utils.spdx.SpdxConstants
 import org.ossreviewtoolkit.utils.spdx.SpdxExpression
 import org.ossreviewtoolkit.utils.spdx.SpdxLicenseWithExceptionExpression
 import org.ossreviewtoolkit.utils.spdx.toSpdx
 import org.ossreviewtoolkit.utils.spdx.toSpdxId
-
-@Serializable
-data class ScanCodeResult(
-    val headers: List<HeaderEntry>,
-    val files: List<FileEntry>,
-    val licenseReferences: List<LicenseReference>? = null // Available only with "--license-references".
-)
-
-@Serializable
-data class HeaderEntry(
-    val toolName: String,
-    val toolVersion: String,
-    val options: Map<String, JsonElement>,
-    val startTimestamp: String,
-    val endTimestamp: String,
-    val outputFormatVersion: String
-) {
-    fun getInput(): File {
-        val inputPath = when (val input = options.getValue("input")) {
-            is JsonPrimitive -> input.content
-            is JsonArray -> input.first().jsonPrimitive.content
-            else -> throw SerializationException("Unknown input element type.")
-        }
-
-        return File(inputPath)
-    }
-
-    fun getPrimitiveOptions(): List<Pair<String, String>> =
-        options.mapNotNull { entry ->
-            (entry.value as? JsonPrimitive)?.let { entry.key to it.content }
-        }
-}
 
 /**
  * An interface to be able to treat all versions of file entries the same.
@@ -148,98 +113,4 @@ sealed interface FileEntry {
 @Serializable
 data class LicenseDetection(
     val matches: List<LicenseEntry>
-)
-
-/**
- * An interface to be able to treat all versions of license entries the same.
- *
- * Note that the data class constructors of the individual version's implementation of this interface should contain
- * only those properties which are actually present in the data, in the order used in the data, to exactly resemble that
- * version's data model. Other properties required to implement the interface should be added to the body of the data
- * class to clearly separate data model properties from "synthetic" interface properties.
- */
-sealed interface LicenseEntry {
-    val licenseExpression: String
-    val startLine: Int
-    val endLine: Int
-    val score: Float
-    val matchedText: String?
-
-    @Serializable
-    data class Version1(
-        val key: String,
-        override val score: Float,
-        val spdxLicenseKey: String? = null, // This might be explicitly set to null in JSON.
-        override val startLine: Int,
-        override val endLine: Int,
-        val matchedRule: LicenseRule,
-        override val matchedText: String? = null
-    ) : LicenseEntry {
-        override val licenseExpression = matchedRule.licenseExpression
-    }
-
-    @Serializable
-    data class Version3(
-        override val licenseExpression: String,
-        val spdxLicenseExpression: String? = null,
-        val fromFile: String? = null,
-        override val startLine: Int,
-        override val endLine: Int,
-        override val score: Float,
-        override val matchedText: String? = null
-    ) : LicenseEntry
-
-    @Serializable
-    data class Version4(
-        override val licenseExpression: String,
-        val licenseExpressionSpdx: String? = null,
-        val fromFile: String? = null,
-        override val startLine: Int,
-        override val endLine: Int,
-        override val score: Float,
-        override val matchedText: String? = null
-    ) : LicenseEntry
-}
-
-@Serializable
-data class LicenseRule(
-    val licenseExpression: String
-)
-
-/**
- * An interface to be able to treat all versions of copyright entries the same.
- *
- * Note that the data class constructors of the individual version's implementation of this interface should contain
- * only those properties which are actually present in the data, in the order used in the data, to exactly resemble that
- * version's data model. Other properties required to implement the interface should be added to the body of the data
- * class to clearly separate data model properties from "synthetic" interface properties.
- */
-sealed interface CopyrightEntry {
-    val statement: String
-    val startLine: Int
-    val endLine: Int
-
-    @Serializable
-    data class Version1(
-        val value: String,
-        override val startLine: Int,
-        override val endLine: Int
-    ) : CopyrightEntry {
-        override val statement = value
-    }
-
-    @Serializable
-    data class Version2(
-        val copyright: String,
-        override val startLine: Int,
-        override val endLine: Int
-    ) : CopyrightEntry {
-        override val statement = copyright
-    }
-}
-
-@Serializable
-data class LicenseReference(
-    val key: String,
-    val spdxLicenseKey: String
 )
